routing: extract keepalive runner from tunnel.go
Second pure code-motion increment on pkg/routing/tunnel.go (#5661 Go
control-plane modularity cohort). The per-tunnel keepalive engine — the
runner goroutine, its lifecycle, the probe loop/tick, and the
KeepaliveState machinery — is a self-contained concern distinct from the
GRE/IPIP/anchor apply reconcile. The prober primitives already live in
tunnel_keepalive.go (#1918); this moves the manager-side runner half.

New pkg/routing/tunnel_keepalive_runner.go holds, byte-identical:
  - KeepaliveState + keepaliveRunner types, matches() identity check
  - keepaliveProber resolver
  - stopAll / stopAllKeepalivesLocked / stopKeepaliveLocked / startKeepalive
  - clampKeepaliveIntervalSec / keepaliveProbeDeadline / maxKeepaliveIntervalSec
  - keepaliveLoop / keepaliveTick / nextSeq
  - clearUnknownLocked / markUnknownLocked / classifyErrnoString
  - GetKeepaliveState

No declarations renamed, no logic changed, same package. tunnel.go drops
from 1795 to 1362 prod LOC — comfortably under the ~2000 modularity
threshold. What remains is the cohesive GRE/IPIP + anchor-TUN reconcile,
VRF-claim, address, Clear and GetStatus surface. go build ./... and
go test ./pkg/routing/... pass; gofmt clean.

Partial progress on #5661 (low-risk routing/snmp/cli files); HA-critical
files (daemon/vrrp/cluster) tracked separately.
